README.md file almost completed

+++ server/services/studyPlan_service.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const { validatePlan } = require("../validate");
+
 
 class StudyPlan_Service {
     
@@ -10,6 +12,11 @@ class StudyPlan_Service {
     }
 
     studyPlan_init = async(req,res) => { 
+
+        if(req.user === undefined) {
+            res.status(401).json("You are not authenticated!").end();
+            return;
+        }
         
         this.#dao.studyPlan_init_DB(req.user.id, req.body).then(() => {  
             return res.status(201).json("Empty Study Plan created!").end();
@@ -17,6 +24,12 @@ class StudyPlan_Service {
     }
 
     get_option = async(req,res) => {   
+
+        if(req.user === undefined) {
+            res.status(401).json("You are not authenticated!").end();
+            return;
+        }
+
         this.#dao.get_option_DB(req.user.id).then((option) => {  
             return res.status(200).json(option).end();
         }).catch((error) => res.status(error).json("Something was wrong!").end());
@@ -24,15 +37,38 @@ class StudyPlan_Service {
 
 
     get_studyPlan = async(req,res) => {   
+
+        if(req.user === undefined) {
+            res.status(401).json("You are not authenticated!").end();
+            return;
+        }
+
         this.#dao.get_studyPlan_DB(req.user.id).then((studyPlan) => {  
             return res.status(200).json(studyPlan).end();
         }).catch((error) => res.status(error).json("Something was wrong!").end());
     }
 
-    post_studyPlan = async(req,res) => { 
-        this.#dao.post_studyPlan_DB(req.user.id, req.body).then(() => {  
-            return res.status(201).json("Added!").end();
-        }).catch((error) => res.status(error).json("Something was wrong!").end());
+    post_studyPlan = async(req,res) => {
+
+        if(req.user === undefined) {
+            res.status(401).json("You are not authenticated!").end();
+            return;
+        }
+        
+        // Validate StudyPlan
+        let valid = await validatePlan(req.body);
+
+        if(valid) {
+            this.#dao.post_studyPlan_DB(req.user.id, req.body).then(() => {  
+                res.status(201).json("Study plan has been created!").end();
+            }).catch((error) => res.status(error).json("Something was wrong!").end());   
+        } else {
+            res.status(422).json("Study Plan is not valid! Cannot be created!").end();
+        }
+
+        // this.#dao.post_studyPlan_DB(req.user.id, req.body).then(() => {  
+        //     return res.status(201).json("Added!").end();
+        // }).catch((error) => res.status(error).json("Something was wrong!").end());
     }
 
     confirm_studyPlan = async(req,res) => {   
@@ -42,6 +78,12 @@ class StudyPlan_Service {
     }
 
     destroy_studyPlan = async(req, res) => {
+
+        if(req.user === undefined) {
+            res.status(401).json("You are not authenticated!").end();
+            return;
+        }
+
         this.#dao.destroy_studyPlan_DB(req.user.id, req.body).then(() => {  
             return res.status(200).json("Destroyed!").end();
         }).catch((error) => res.status(error).json("Something was wrong!").end());
@@ -52,6 +94,8 @@ class StudyPlan_Service {
     //         return res.status(204).json("Course remove from the Study Plan!").end();
     //     }).catch((error) => res.status(error).json("Something was wrong!").end());
     // }
+
+    
 
 
 }
